Changes to comply more precisely with the spec for connectivity checks from the client side
git-svn-id: http://svn.littleshoot.org/svn/littleshoot/trunk/common/ice@1108 8d7aff80-242d-0410-bcee-ed6f15385b3d

+++ src/main/java/org/lastbamboo/common/ice/candidate/IceTcpHostPassiveCandidate.java
@@ -35,7 +35,7 @@ public class IceTcpHostPassiveCandidate extends AbstractIceCandidate
      * @param componentId The component ID.
      */
     public IceTcpHostPassiveCandidate(final InetSocketAddress socketAddress, 
-        final String foundation, final boolean controlling, final int priority, 
+        final String foundation, final boolean controlling, final long priority, 
         final int componentId)
         {
         super(socketAddress, foundation,
